Pull almost all of the logic out of `main-client`.
...into a new `top-client` module. This paves the way for new-style
configuration of the client.
