+++ local-modules/@bayou/main-client/README.md
@@ -1,10 +1,7 @@
 @bayou/main-client
 ==================
 
-This module contains the top level code for the client, including a
-`package.json` which is specific to the client side. That is, the modules
-referenced here become available for use in the browser, but are not included on
-the server side, at least not due to being referenced here.
+This module contains code to run a client in the default configuration.
 
 - - - - - - - - - -
