+++ local-modules/@bayou/main-client/package.json
@@ -6,12 +6,7 @@
   "testMain": "tests/index.js",
 
   "dependencies": {
-    "@bayou/api-common": "local",
-    "@bayou/app-common": "local",
     "@bayou/config-common-default": "local",
-    "@bayou/doc-client": "local",
-    "@bayou/env-client": "local",
-    "@bayou/see-all-client": "local",
-    "@bayou/testing-client": "local"
+    "@bayou/top-client": "local"
   }
 }
