+++ local-modules/@bayou/top-client/index.js
@@ -1,0 +1,7 @@
+// Copyright 2016-2018 the Bayou Authors (Dan Bornstein et alia).
+// Licensed AS IS and WITHOUT WARRANTY under the Apache License,
+// Version 2.0. Details: <http://www.apache.org/licenses/LICENSE-2.0>
+
+import Client from './Client';
+
+export { Client };
